added UserServie and test files, moved setCurrentUser() and getCurrentUser() to UserService.
