+++ src/app/home-page/home-page.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 //TODO Remove these 2 imports
-import { LoginService } from './../login/login.service';
-import { User } from './../user/user'
+import { UserService } from './../_services/user.service'
 
 @Component({
   selector: 'app-home-page',
@@ -11,13 +10,13 @@ import { User } from './../user/user'
 export class HomePageComponent implements OnInit {
 
 // TODO remmove loginservice below
-  constructor( private loginService: LoginService) { }
+  constructor( private userService: UserService) { }
 
   ngOnInit() { }
 
   // TODO: Remove these lines when done, it is just used to verify data capturing to correct variable
   verifyGetCurrentUser($event){
-    let currentUser = this.loginService.getCurrentUser();
+    let currentUser = this.userService.getCurrentUser();
     console.log( "In HomePageComponent: getCurrentUser is: " + JSON.stringify(currentUser));
   }
