+++ src/app/login/login.service.ts
@@ -8,14 +8,7 @@ import { User, Role } from '../user/user';
 })
 export class LoginService {
 
-  private currentRole: Role;
-  private currentUser: User;
-
-  constructor(private apiService: ApiService) {
-  //  this.currentRole = new Role(0, '');
-//    this.currentUser= new User(0, this.currentRole, '', '', 1, '', '', '');
-  //  this.currentRole = new Role(0, undefined);
-  //  this.currentUser= new User(0, this.currentRole, undefined, undefined, 1, undefined, undefined, undefined);
+  constructor( private apiService: ApiService ) {
   }
 
   // method for login an existing user
@@ -26,17 +19,6 @@ export class LoginService {
 
     return this.apiService.get(apiVerifyCredentialsUrl, username, password);
 
-  }
-
-
-  // method for saving user data returned from user login()
-  public setCurrentUser( newUser: User ) {
-    this.currentUser = newUser;
-  }
-
-  // method for getting the current user logged in
-  public getCurrentUser(): User {
-    return this.currentUser;
   }
 
 }
